rename *Worker to *Runtime, and create AgentRuntimeBase. Align runtime interface with python
