+++ dotnet/src/Microsoft.AutoGen/Core/AgentExtensions.cs
@@ -22,7 +22,7 @@ public static class AgentExtensions
     public static Activity? ExtractActivity(this Agent agent, string activityName, IDictionary<string, string> metadata)
     {
         Activity? activity;
-        (string? traceParent, string? traceState) = IAgentWorkerExtensions.GetTraceIdAndState(agent.Worker, metadata);
+        (string? traceParent, string? traceState) = IAgentRuntimeExtensions.GetTraceIdAndState(agent.Worker, metadata);
         if (!string.IsNullOrEmpty(traceParent))
         {
             if (ActivityContext.TryParse(traceParent, traceState, isRemote: true, out var parentContext))
@@ -43,7 +43,7 @@ public static class AgentExtensions
                     activity.TraceStateString = traceState;
                 }
 
-                var baggage = IAgentWorkerExtensions.ExtractMetadata(agent.Worker, metadata);
+                var baggage = IAgentRuntimeExtensions.ExtractMetadata(agent.Worker, metadata);
 
                 foreach (var baggageItem in baggage)
                 {
